#292 [refactor] Withdraw 로직 리팩토링
