+++ app/src/main/java/hous/release/android/presentation/settings/SettingsActivity.kt
@@ -2,7 +2,6 @@ package hous.release.android.presentation.settings
 
 import android.content.Intent
 import android.net.Uri
-import android.os.Build
 import android.os.Bundle
 import androidx.activity.viewModels
 import dagger.hilt.android.AndroidEntryPoint
@@ -71,11 +70,7 @@ class SettingsActivity : BindingActivity<ActivitySettingsBinding>(R.layout.activ
 
     private fun initFeedbackClickListener() {
         binding.tvSettingsFeedback.setOnClickListener {
-            startActivity(
-                Intent(this, WithdrawActivity::class.java).apply {
-                    putExtra(IS_DELETING, false)
-                }
-            )
+            // TODO by 이빵주 : FeedbackActivity 만들어서 띄우기
         }
     }
 
@@ -88,9 +83,7 @@ class SettingsActivity : BindingActivity<ActivitySettingsBinding>(R.layout.activ
     private fun initWithdrawClickListener() {
         binding.tvSettingsWithdraw.setOnClickListener {
             startActivity(
-                Intent(this, WithdrawActivity::class.java).apply {
-                    putExtra(IS_DELETING, true)
-                }
+                Intent(this, WithdrawActivity::class.java)
             )
         }
     }
@@ -143,27 +136,7 @@ class SettingsActivity : BindingActivity<ActivitySettingsBinding>(R.layout.activ
         }.show(supportFragmentManager, WarningDialogFragment.DIALOG_WARNING)
     }
 
-    private fun getDeviceName(): String {
-        val manufacturer = Build.MANUFACTURER
-        val model = Build.MODEL
-        return if (model.startsWith(manufacturer)) {
-            capitalize(model)
-        } else {
-            "${capitalize(manufacturer)} $model"
-        }
-    }
-
-    private fun capitalize(str: String?): String {
-        return if (str.isNullOrEmpty()) {
-            getString(R.string.settings_feedback_empty_string)
-        } else {
-            val first = str[0]
-            Character.toUpperCase(first).toString() + str.substring(1)
-        }
-    }
-
     companion object {
         const val HAS_ROOM = "HAS_ROOM"
-        const val IS_DELETING = "isDeleting"
     }
 }
