+++ app/src/main/java/hous/release/android/presentation/withdraw/withdraw/WithdrawFragment.kt
@@ -7,8 +7,6 @@ import androidx.navigation.fragment.findNavController
 import dagger.hilt.android.AndroidEntryPoint
 import hous.release.android.R
 import hous.release.android.databinding.FragmentWithdrawBinding
-import hous.release.android.presentation.withdraw.WithdrawActivity.Companion.LOCATION
-import hous.release.android.presentation.withdraw.WithdrawActivity.Companion.WITHDRAW
 import hous.release.android.util.UiEvent
 import hous.release.android.util.binding.BindingFragment
 import hous.release.android.util.dialog.LoadingDialogFragment
@@ -17,7 +15,10 @@ import hous.release.android.util.extension.repeatOnStarted
 @AndroidEntryPoint
 class WithdrawFragment : BindingFragment<FragmentWithdrawBinding>(R.layout.fragment_withdraw) {
     private val withdrawViewModel by viewModels<WithdrawViewModel>()
-    private val bundle = Bundle()
+
+    private val loadingDialogFragment by lazy {
+        childFragmentManager.findFragmentByTag(LoadingDialogFragment.TAG) as? LoadingDialogFragment
+    }
 
     override fun onViewCreated(view: View, savedInstanceState: Bundle?) {
         super.onViewCreated(view, savedInstanceState)
@@ -29,9 +30,6 @@ class WithdrawFragment : BindingFragment<FragmentWithdrawBinding>(R.layout.fragm
     private fun collectUiEvent() {
         repeatOnStarted {
             withdrawViewModel.uiEvent.collect { uiEvent ->
-                val loadingDialogFragment =
-                    childFragmentManager.findFragmentByTag(LoadingDialogFragment.TAG) as? LoadingDialogFragment
-
                 when (uiEvent) {
                     UiEvent.LOADING -> {
                         loadingDialogFragment?.show(
@@ -40,12 +38,7 @@ class WithdrawFragment : BindingFragment<FragmentWithdrawBinding>(R.layout.fragm
                     }
                     UiEvent.SUCCESS -> {
                         loadingDialogFragment?.dismiss()
-                        findNavController().navigate(
-                            R.id.action_withdrawFragment_to_withdrawDoneFragment,
-                            bundle.apply {
-                                putString(LOCATION, WITHDRAW)
-                            }
-                        )
+                        findNavController().navigate(R.id.action_withdrawFragment_to_withdrawDoneFragment)
                     }
                     UiEvent.ERROR -> {
                         loadingDialogFragment?.dismiss()
