+++ app/src/main/res/layout/fragment_withdraw_done.xml
@@ -5,9 +5,6 @@
 
     <data>
 
-        <variable
-            name="viewModel"
-            type="hous.release.android.presentation.withdraw.withdrawDone.WithdrawDoneViewModel" />
     </data>
 
     <androidx.constraintlayout.widget.ConstraintLayout
@@ -23,36 +20,33 @@
             android:layout_marginEnd="14dp"
             android:clickable="true"
             android:padding="10dp"
-            android:text="@{viewModel.layout == `withdraw` ? @string/withdraw_done_return : @string/feedback_done_return}"
+            android:text="@string/withdraw_done_return"
             android:textColor="@color/hous_blue"
             android:theme="@style/B2"
             app:layout_constraintEnd_toEndOf="parent"
-            app:layout_constraintTop_toTopOf="parent"
-            tools:text="@string/withdraw_done_return" />
+            app:layout_constraintTop_toTopOf="parent" />
 
         <ImageView
             android:id="@+id/iv_withdraw_done_graphic"
             android:layout_width="wrap_content"
             android:layout_height="wrap_content"
             android:layout_marginTop="208dp"
-            android:src="@{viewModel.layout == `withdraw` ? @drawable/ic_withdraw_done : @drawable/ic_feedback_done}"
+            android:src="@drawable/ic_withdraw_done"
             app:layout_constraintEnd_toEndOf="parent"
             app:layout_constraintStart_toStartOf="parent"
-            app:layout_constraintTop_toBottomOf="@id/tv_withdraw_done_return"
-            tools:src="@drawable/ic_feedback_done" />
+            app:layout_constraintTop_toBottomOf="@id/tv_withdraw_done_return" />
 
         <TextView
             android:id="@+id/tv_withdraw_done_title"
             android:layout_width="wrap_content"
             android:layout_height="wrap_content"
             android:layout_marginTop="20dp"
-            android:text="@{viewModel.layout == `withdraw` ? @string/withdraw_done_title : @string/feedback_done_title}"
+            android:text="@string/withdraw_done_title"
             android:textColor="@color/hous_black"
             android:theme="@style/H3"
             app:layout_constraintEnd_toEndOf="parent"
             app:layout_constraintStart_toStartOf="parent"
-            app:layout_constraintTop_toBottomOf="@id/iv_withdraw_done_graphic"
-            tools:text="@string/withdraw_done_title" />
+            app:layout_constraintTop_toBottomOf="@id/iv_withdraw_done_graphic" />
 
         <TextView
             android:id="@+id/tv_withdraw_done_desc"
@@ -60,13 +54,12 @@
             android:layout_height="wrap_content"
             android:layout_marginTop="8dp"
             android:gravity="center"
-            android:text="@{viewModel.layout == `withdraw` ? @string/withdraw_done_desc : @string/feedback_done_desc}"
+            android:text="@string/withdraw_done_desc"
             android:textColor="@color/hous_g_6"
             android:theme="@style/B1"
             app:layout_constraintEnd_toEndOf="parent"
             app:layout_constraintStart_toStartOf="parent"
-            app:layout_constraintTop_toBottomOf="@id/tv_withdraw_done_title"
-            tools:text="@string/withdraw_done_desc" />
+            app:layout_constraintTop_toBottomOf="@id/tv_withdraw_done_title" />
 
     </androidx.constraintlayout.widget.ConstraintLayout>
 </layout>
